Put UserProblemQuestions in UserStats as a map.
Because this should reduce the number of reads we need to do just to
find a suitable next question. And because we then only need to
write one entity, not two, each time a question is answered. Apparently,
it doesn't matter how big the entity is, but how much of it has changed.