Presenters: Always scroll to top when revealing.
By deriving from a shared base class the handles the event.
Otherwise, we show new page content but the browser window doesn't take
us to the top, like it would with a real link to a real different page.
This doesn't seem like something I should have to hack together.