UserStats: getTopProblemQuestionHistories(): Expect a manual update.
Via a new calcTopProblemQuestionHistories() method, which only needs
to be called by the server-side code.

This avoids unnecessary work on the client side, and lets the JSON
serialization/deserialization work without having to know about
(and  therefore have to transmit) the larger set of data that this
"top" list is based on.

Now the section history sidebar actually shows the problem questions.
This was a regression caused by the port to use a REST/JSON API.