UserResource.get(): Don't take a request-url.
And don't provide it from the client side.

Because all logins are now from the login page, and we don't want to
return to that, so we will instead always return to the home page.

Returning to the previous page makes sense if we can login directly from
any page.